# proto/faiss_def.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "faiss_def.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/faiss/1.0/hsearch": {
      "post": {
        "operationId": "HSearch",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/faiss_serverHSearchResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/faiss_serverHSearchRequest"
            }
          }
        ],
        "tags": [
          "FaissService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "HSearchRequestDistanceType": {
      "type": "string",
      "enum": [
        "Euclid",
        "Cosine"
      ],
      "default": "Euclid"
    },
    "HSearchResponseResult": {
      "type": "object",
      "properties": {
        "score": {
          "type": "number",
          "format": "float"
        },
        "id": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
...
    "faiss_serverHSearchRequest": {
      "type": "object",
      "properties": {
        "db_name": {
          "type": "string"
        },
        "feature": {
          "type": "string",
          "format": "byte"
        },
        "top_k": {
          "type": "string",
          "format": "uint64"
        },
        "distance_type": {
          "$ref": "#/definitions/HSearchRequestDistanceType"
        },
        "request_id": {
          "type": "string"
        }
      },
      "title": "ANN检索请求"
    },
    "faiss_serverHSearchResponse": {
      "type": "object",
      "properties": {
        "results": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/HSearchResponseResult"
          }
        },
        "request_id": {
          "type": "string"
        },
        "error_code": {
          "type": "string",
          "format": "int64"
        },
        "error_msg": {
          "type": "string"
        }
      },
      "title": "ANN 检索返回"
    },
...
  }
}